chore: resolved sonar issues
Refs: #207

+++ src/main/java/ch/sbb/polarion/extension/diff_tool/service/MergeService.java
@@ -223,10 +223,13 @@ public class MergeService {
                 IWorkItem pairedWorkItem = polarionService.getPairedWorkItems(source, context.getTargetModule().getProjectId(), context.linkRole).stream().findFirst().orElse(null);
                 if (pairedWorkItem != null) {
                     newWorkItem = insertWorkItem(source, context, false);
-                    context.reportEntry(CREATED, pair, "workitem '%s' moved from '%s'".formatted(newWorkItem.getId(), pairedWorkItem.getModule() != null ? pairedWorkItem.getModule().getTitleWithSpace() : "tracker"));
+                    if (newWorkItem != null) {
+                        context.reportEntry(CREATED, pair, "workitem '%s' moved from '%s'".formatted(newWorkItem.getId(), pairedWorkItem.getModule() != null ? pairedWorkItem.getModule().getTitleWithSpace() : "tracker"));
+                    } else {
+                        context.reportEntry(CREATION_FAILED, pair, "new workitem based on source workitem '%s' NOT created".formatted(pairedWorkItem.getModule() != null ? pairedWorkItem.getModule().getTitleWithSpace() : "tracker"));
+                    }
                 } else {
                     newWorkItem = copyWorkItemToDocument(source, (InternalWriteTransaction) transaction, context, pair);
-                    context.reportEntry(CREATED, pair, "new workitem '%s' based on source workitem '%s' created".formatted(newWorkItem.getId(), source.getId()));
                 }
                 context.bindCounterpartItem(pair,
                         WorkItem.of(newWorkItem,
@@ -262,8 +265,10 @@ public class MergeService {
     private boolean deleteItem(WorkItemsPair pair, WorkItemsMergeContext context) {
         if (context.getSourceWorkItem(pair) == null && context.getTargetWorkItem(pair) != null) {
             IWorkItem target = getWorkItem(context.getTargetWorkItem(pair));
-            target.delete();
-            return true;
+            if (target != null) {
+                target.delete();
+                return true;
+            }
         }
         return false;
     }
@@ -711,7 +716,7 @@ public class MergeService {
             }
         }
     }
-
+    @Nullable
     private IWorkItem getWorkItem(@Nullable WorkItem workItem) {
         return workItem != null ? polarionService.getWorkItem(workItem.getProjectId(), workItem.getId(), workItem.getRevision()) : null;
     }
